+++ Frends.Avro.Deserialize/Frends.Avro.Deserialize/Definitions/Input.cs
@@ -1,13 +1,17 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace Frends.Avro.Deserialize.Definitions;
 
 /// <summary>
-/// Input parameters.
+/// Input parameters for Avro deserialization operation.
 /// </summary>
 public class Input
 {
     /// <summary>
-    /// Path to the Avro file you want to deserialize
+    /// Gets or sets the path to the Avro file you want to deserialize.
+    /// The file must be a valid Avro format file.
     /// </summary>
     /// <example>C:\results\myfile.avro</example>
-    public string AvroFilePath { get; init; }
+    [DisplayFormat(DataFormatString = "Text")]
+    public string FilePath { get; init; }
 }