+++ Frends.Avro.Deserialize/Frends.Avro.Deserialize/Definitions/Options.cs
@@ -1,0 +1,27 @@
+using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
+
+namespace Frends.Avro.Deserialize.Definitions;
+
+/// <summary>
+/// Configuration options for Avro deserialization operation.
+/// </summary>
+public class Options
+{
+    /// <summary>
+    /// Gets or sets a value indicating whether to throw an exception when deserialization fails.
+    /// When set to true, exceptions will be thrown on failure. When set to false, errors will be returned in the Result object.
+    /// </summary>
+    /// <example>true</example>
+    [DefaultValue(true)]
+    public bool ThrowErrorOnFailure { get; set; } = true;
+
+    /// <summary>
+    /// Gets or sets a custom error message to use when deserialization fails and ThrowErrorOnFailure is false.
+    /// If empty, the original exception message will be used.
+    /// </summary>
+    /// <example>Custom error occurred during Avro deserialization</example>
+    [DefaultValue("")]
+    [DisplayFormat(DataFormatString = "Text")]
+    public string ErrorMessageOnFailure { get; set; } = "";
+}